user permissions in controllers

+++ app/Http/Controllers/CarBodyTypeController.php
@@ -4,15 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\CarBodyType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use function Symfony\Component\String\upper;
+use const http\Client\Curl\AUTH_ANY;
 
 class CarBodyTypeController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function index()
     {
@@ -23,11 +25,14 @@ class CarBodyTypeController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function create()
     {
-        return view('car_body_types.create');
+        if (Auth::user() && Auth::user()->can('create', CarBodyType::class))
+            return view('car_body_types.create');
+        else
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Tenant has been edited successfully');
     }
 
     /**
@@ -38,14 +43,18 @@ class CarBodyTypeController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-        ]);
-        $type = new CarBodyType();
-        $type->name = strtoupper($request->name);
-        $type->timestamps = false;
-        $type->save();
-        return redirect()->route('carBodyType.index')->with('successMsg', 'Type has been created successfully');
+        if (Auth::user() && Auth::user()->can('create', CarBodyType::class)) {
+            $request->validate([
+                'name' => 'required',
+            ]);
+            $type = new CarBodyType();
+            $type->name = strtoupper($request->name);
+            $type->timestamps = false;
+            $type->save();
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Type has been created successfully');
+        } else {
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Tenant has been edited successfully');
+        }
     }
 
     /**
@@ -56,14 +65,14 @@ class CarBodyTypeController extends Controller
      */
     public function show(CarBodyType $carBodyType)
     {
-        $transports=$carBodyType->transports;
+        $transports = $carBodyType->transports;
         $top5 = [];
         if (count($transports) < 6) {
-            $top5=$transports;
+            $top5 = $transports;
         } else {
-            $numbers=array_rand(range(0,count($transports)-1),5);
-            foreach ($numbers as $num){
-                array_push($top5,$transports[$num]);
+            $numbers = array_rand(range(0, count($transports) - 1), 5);
+            foreach ($numbers as $num) {
+                array_push($top5, $transports[$num]);
             }
         }
         return view('car_body_types.show', compact('carBodyType'), compact('top5'));
@@ -77,7 +86,11 @@ class CarBodyTypeController extends Controller
      */
     public function edit(CarBodyType $carBodyType)
     {
-        return view('car_body_types.edit', compact('carBodyType'));
+        if (Auth::user() && Auth::user()->can('edit', CarBodyType::class)) {
+            return view('car_body_types.edit', compact('carBodyType'));
+        } else {
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Tenant has been edited successfully');
+        }
     }
 
     /**
@@ -89,14 +102,18 @@ class CarBodyTypeController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required',
-        ]);
-        $type = CarBodyType::find($id);
-        $type->name = strtoupper($request->name);
-        $type->timestamps = false;
-        $type->save();
-        return redirect()->route('carBodyType.index')->with('successMsg', 'Type has been edited successfully');
+        if (Auth::user() && Auth::user()->can('edit', CarBodyType::class)) {
+            $request->validate([
+                'name' => 'required',
+            ]);
+            $type = CarBodyType::find($id);
+            $type->name = strtoupper($request->name);
+            $type->timestamps = false;
+            $type->save();
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Type has been edited successfully');
+        } else {
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Tenant has been edited successfully');
+        }
     }
 
     /**
@@ -107,7 +124,11 @@ class CarBodyTypeController extends Controller
      */
     public function destroy(CarBodyType $carBodyType)
     {
-        $carBodyType->delete();
-        return redirect()->route('carBodyType.index')->with('successMsg', 'Type has been deleted successfully');
+        if (Auth::user() && Auth::user()->can('delete', CarBodyType::class)) {
+            $carBodyType->delete();
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Type has been deleted successfully');
+        } else {
+            return redirect()->route('carBodyType.index')->with('successMsg', 'Tenant has been edited successfully');
+        }
     }
 }
