+++ app/Http/Controllers/RentController.php
@@ -6,7 +6,9 @@ use App\Models\Owner;
 use App\Models\Rent;
 use App\Models\Tenant;
 use App\Models\Transport;
+use App\Providers\RouteServiceProvider;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class RentController extends Controller
@@ -26,9 +28,13 @@ class RentController extends Controller
      */
     public function index()
     {
-        $owners = Owner::all();
-        $rents = Rent::paginate(20);
-        return view('rents.index', ['owners' => $owners, 'rents' => $rents]);
+        if (Auth::user() && Auth::user()->can('view', Rent::class)) {
+            $owners = Owner::all();
+            $rents = Rent::paginate(20);
+            return view('rents.index', ['owners' => $owners, 'rents' => $rents]);
+        } else {
+            return redirect()->route(RouteServiceProvider::HOME)->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -38,16 +44,20 @@ class RentController extends Controller
      */
     public function create($transportId)
     {
-        $owners = DB::select(DB::raw('select owners.id, owners.name
+        if (Auth::user() && Auth::user()->can('create', Rent::class)) {
+            $owners = DB::select(DB::raw('select owners.id, owners.name
                                                 from owners inner join transports on transports.owner_id=owners.id
                                                 where transports.id=' . $transportId));
-        $tenants = Tenant::all();
-        $transports = DB::table('transports')->where('id', '=', $transportId)->get();
-        return view('rents.create', [
-            'owners' => $owners,
-            'tenants' => $tenants,
-            'transports' => $transports,
-            'transportId' => $transportId]);
+            $tenants = Tenant::all();
+            $transports = DB::table('transports')->where('id', '=', $transportId)->get();
+            return view('rents.create', [
+                'owners' => $owners,
+                'tenants' => $tenants,
+                'transports' => $transports,
+                'transportId' => $transportId]);
+        } else {
+            return redirect()->route('rent.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -58,21 +68,25 @@ class RentController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate(self::VALIDATION_RULE);
-        $tenant = DB::select("select top 1 * from tenants where name='$request->tenant'");
-        $transport = Transport::find($request->id_transport);
-        $rent = new Rent();
-        $rent->date = $request->date;
-        $rent->id_transport = $request->id_transport;
-        $rent->id_tenant = $tenant[0]->id;
-        $rent->rental_period = $request->rental_period;
-        $rent->id_owner = $request->id_owner;
-        $rent->timestamps = false;
-        $rent->save();
-        $transport->rental_times += 1;
-        $transport->timestamps = false;
-        $transport->save();
-        return redirect()->route('rent.index')->with('successMsg', 'Rent has been created successfully');
+        if (Auth::user() && Auth::user()->can('create', Rent::class)) {
+            $request->validate(self::VALIDATION_RULE);
+            $tenant = DB::select("select top 1 * from tenants where name='$request->tenant'");
+            $transport = Transport::find($request->id_transport);
+            $rent = new Rent();
+            $rent->date = $request->date;
+            $rent->id_transport = $request->id_transport;
+            $rent->id_tenant = $tenant[0]->id;
+            $rent->rental_period = $request->rental_period;
+            $rent->id_owner = $request->id_owner;
+            $rent->timestamps = false;
+            $rent->save();
+            $transport->rental_times += 1;
+            $transport->timestamps = false;
+            $transport->save();
+            return redirect()->route('rent.index')->with('successMsg', 'Rent has been created successfully');
+        } else {
+            return redirect()->route('rent.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -83,7 +97,10 @@ class RentController extends Controller
      */
     public function show(Rent $rent)
     {
-        return view('rents.show', compact('rent'));
+        if (Auth::user() && Auth::user()->can('view', Rent::class))
+            return view('rents.show', compact('rent'));
+        else
+            return redirect()->route('rent.index')->with('fail', "You dont have permission");
     }
 
     /**
@@ -94,15 +111,19 @@ class RentController extends Controller
      */
     public function edit(Rent $rent)
     {
-        $owners = DB::select(DB::raw('select owners.id, owners.name
+        if (Auth::user() && Auth::user()->can('create', Rent::class)) {
+            $owners = DB::select(DB::raw('select owners.id, owners.name
                                                 from owners inner join transports on transports.owner_id=owners.id
                                                 where transports.id=' . $rent->id_transport));
-        $tenants = Tenant::all();
-        $transports = DB::table('transports')->where('id', '=', $rent->id_transport)->get();
-        return view('rents.edit', compact('rent'), [
-            'owners' => $owners,
-            'tenants' => $tenants,
-            'transports' => $transports]);
+            $tenants = Tenant::all();
+            $transports = DB::table('transports')->where('id', '=', $rent->id_transport)->get();
+            return view('rents.edit', compact('rent'), [
+                'owners' => $owners,
+                'tenants' => $tenants,
+                'transports' => $transports]);
+        } else {
+            return redirect()->route('rent.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -114,16 +135,20 @@ class RentController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate(self::VALIDATION_RULE);
-        $rent = Rent::find($id);
-        $rent->date = $request->date;
-        $rent->id_transport = $request->id_transport;
-        $rent->id_tenant = $request->id_tenant;
-        $rent->rental_period = $request->rental_period;
-        $rent->id_owner = $request->id_owner;
-        $rent->timestamps = false;
-        $rent->save();
-        return redirect()->route('rent.index')->with('successMsg', 'Rent has been edited successfully');
+        if (Auth::user() && Auth::user()->can('create', Rent::class)) {
+            $request->validate(self::VALIDATION_RULE);
+            $rent = Rent::find($id);
+            $rent->date = $request->date;
+            $rent->id_transport = $request->id_transport;
+            $rent->id_tenant = $request->id_tenant;
+            $rent->rental_period = $request->rental_period;
+            $rent->id_owner = $request->id_owner;
+            $rent->timestamps = false;
+            $rent->save();
+            return redirect()->route('rent.index')->with('successMsg', 'Rent has been edited successfully');
+        } else {
+            return redirect()->route('rent.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -134,11 +159,16 @@ class RentController extends Controller
      */
     public function destroy(Rent $rent)
     {
-        $rent->delete();
-        return redirect()->route('rent.index')->with('successMsg', 'Rent has been deleted successfully');
+        if (Auth::user() && Auth::user()->can('create', Rent::class)) {
+            $rent->delete();
+            return redirect()->route('rent.index')->with('successMsg', 'Rent has been deleted successfully');
+        } else {
+            return redirect()->route('rent.index')->with('fail', "You dont have permission");
+        }
     }
 
-    public function getModel($id){
+    public function getModel($id)
+    {
         return \App\Models\Transport::find($id)->model;
     }
 }
