+++ app/Http/Controllers/TenantController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Tenant;
+use App\Models\User;
+use App\Providers\RouteServiceProvider;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class TenantController extends Controller
 {
@@ -14,8 +17,12 @@ class TenantController extends Controller
      */
     public function index()
     {
-        $data['tenants'] = Tenant::paginate(20);
-        return view('tenants.index', $data);
+        if (Auth::user() && Auth::user()->can('view', Tenant::class)) {
+            $data['tenants'] = Tenant::paginate(20);
+            return view('tenants.index', $data);
+        } else {
+            return redirect()->route(RouteServiceProvider::HOME)->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -25,27 +32,37 @@ class TenantController extends Controller
      */
     public function create()
     {
-        return view('tenants.create');
+        if (Auth::user() && Auth::user()->can('create', Tenant::class)) {
+            $data['users'] = User::all();
+            return view('tenants.create', $data);
+        } else {
+            return redirect()->route('tenant.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'legal_entity' => 'integer'
-        ]);
-        $tenant = new Tenant();
-        $tenant->name = $request->name;
-        $tenant->legal_entity = $request->legal_entity;
-        $tenant->timestamps = false;
-        $tenant->save();
-        return redirect()->route('tenant.index')->with('successMsg', 'Tenant has been created successfully');
+        if (Auth::user() && Auth::user()->can('create', Tenant::class)) {
+            $request->validate([
+                'name' => 'required',
+                'legal_entity' => 'integer'
+            ]);
+            $tenant = new Tenant();
+            $tenant->name = $request->name;
+            $tenant->legal_entity = $request->legal_entity;
+            $tenant->user_id = $request->user;
+            $tenant->timestamps = false;
+            $tenant->save();
+            return redirect()->route('tenant.index')->with('successMsg', 'Tenant has been created successfully');
+        } else {
+            return redirect()->route('tenant.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -56,8 +73,12 @@ class TenantController extends Controller
      */
     public function show(Tenant $tenant)
     {
-        $rents=$tenant->rents;
-        return view('tenants.show',compact('tenant'),compact('rents'));
+        if (Auth::user() && Auth::user()->can('create', Tenant::class)) {
+            $rents = $tenant->rents;
+            return view('tenants.show', compact('tenant'), compact('rents'));
+        } else {
+            return redirect()->route('tenant.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -68,7 +89,12 @@ class TenantController extends Controller
      */
     public function edit(Tenant $tenant)
     {
-        return view('tenants.edit', compact('tenant'));
+        if (Auth::user() && Auth::user()->can('edit', Tenant::class)) {
+            $users = User::all();
+            return view('tenants.edit', compact('tenant'), compact('users'));
+        } else {
+            return redirect()->route('tenant.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
@@ -76,31 +102,40 @@ class TenantController extends Controller
      *
      * @param \Illuminate\Http\Request $request
      * @param \App\Models\Tenant $tenant
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required',
-            'legal_entity' => 'integer'
-        ]);
-        $tenant = Tenant::find($id);
-        $tenant->name = $request->name;
-        $tenant->legal_entity = $request->legal_entity;
-        $tenant->timestamps = false;
-        $tenant->save();
-        return redirect()->route('tenant.index')->with('successMsg', 'Tenant has been edited successfully');
+        if (Auth::user() && Auth::user()->can('edit', Tenant::class)) {
+            $request->validate([
+                'name' => 'required',
+                'legal_entity' => 'integer'
+            ]);
+            $tenant = Tenant::find($id);
+            $tenant->name = $request->name;
+            $tenant->legal_entity = $request->legal_entity;
+            $tenant->user_id = $request->user;
+            $tenant->timestamps = false;
+            $tenant->save();
+            return redirect()->route('tenant.index')->with('successMsg', 'Tenant has been edited successfully');
+        } else {
+            return redirect()->route('tenant.index')->with('fail', "You dont have permission");
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param \App\Models\Tenant $tenant
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Tenant $tenant)
     {
-        $tenant->delete();
-        return redirect()->route('tenant.index');
+        if (Auth::user() && Auth::user()->can('edit', Tenant::class)) {
+            $tenant->delete();
+            return redirect()->route('tenant.index')->with('successMsg', 'Tenant has been deleted successfully');
+        } else {
+            return redirect()->route('tenant.index')->with('fail', "You dont have permission");
+        }
     }
 }
